+++ database/migrations/2020_02_18_235710_add_relationships_to_subscriptions_table.php
@@ -16,11 +16,11 @@ class AddRelationshipsToSubscriptionsTable extends Migration
         Schema::table('subscriptions', function (Blueprint $table) {
             //
             if (!Schema::hasColumn('subscriptions', 'account_id')) {
-                $table->integer('account_id')->unsigned();
+                $table->bigInteger('account_id')->unsigned();
                 $table->foreign('account_id')->references('id')->on('accounts')->onDelete('cascade');
             }
             if (!Schema::hasColumn('subscriptions', 'subscription_type_id')) {
-                $table->integer('subscription_type_id')->unsigned();
+                $table->bigInteger('subscription_type_id')->unsigned();
                 $table->foreign('subscription_type_id')->references('id')->on('subscription_types')->onDelete('set null');
             }
         });
